test(server): pin every in-process Cli parse against ambient env
#190 pinned http_transport_wiremock.rs by parsing through a clap command
with every `env` fallback reset. Every other in-process `Cli::parse_from`
still inherited them and hand-cleared api_key/api_key_file instead — the
named-field list whose drift #190 was about.

The pin moves to tests/common/pinned_cli.rs, `#[path]`-included by the six
wiremock harnesses and, through lib.rs, by the crate's own unit tests. Not
common/mod.rs: that compiles into every binary saying `mod common;`, where
a helper only some of them use is dead_code under -D warnings (#167).
Nothing in the file names the crate — `bugwarden::` does not resolve
inside `bugwarden` — so the type arrives as a `CommandFactory` parameter
and src/ can include the same file. #190's two self-checks become shared
functions, called from one `#[test]` per pinning binary so each still
proves the property its own harness claims.

Not nil at HEAD, contrary to the issue: an ambient MCP_ALLOWED_HOSTS turns
Host validation on for http_auth_wiremock's serve_guarded and 403s 6 of
its 12 tests, MCP_API_KEY_HEADER fails config's startup-log test, and
BUGZILLA_API_KEY_FILE fails the one server.rs test carrying no manual
clobber. Pinned, that environment changes nothing; with pin_environment
neutered it fails 226 tests — the clobbers are gone — and the self-check
dies in all seven binaries with no hostile environment at all.

config.rs's tests are pinned too, though the issue lists only server.rs:
one of those three live failures was there. env_config.rs stays as it is,
being the test of the fallbacks themselves.
