+++ crates/bugwarden/tests/http_transport_wiremock.rs
@@ -67,7 +67,6 @@ use bugwarden::server::{BugWarden, USER_AGENT};
 use bugwarden_core::client::BugzillaClient;
 use bugwarden_core::guard::Guard;
 use bugwarden_core::policy::Policy;
-use clap::FromArgMatches as _;
 use rmcp::model::{CallToolRequestParams, CallToolResult};
 use rmcp::service::{RoleClient, RunningService};
 use rmcp::transport::streamable_http_client::StreamableHttpClientTransportConfig;
@@ -81,35 +80,15 @@ use wiremock::{Mock, MockServer, ResponseTemplate};
 #[path = "common/raw_post.rs"]
 mod raw_post;
 
-/// Drop every `env` fallback from `cmd`, so parsing it consults only argv.
-///
-/// `Arg::env` snapshots the variable when the command is built, so the
-/// reset covers fields added to `Cli` after this line was written — which
-/// is the point. The sibling `AMBIENT_VARS` list in `http_auth_wiremock.rs`
-/// solves the other half of the same problem and is deliberately not
-/// shared: it scrubs the environment of a *spawned* binary, and doing that
-/// in-process would race every other test in this binary (`env_config.rs`).
-fn pin_environment(cmd: clap::Command) -> clap::Command {
-    cmd.mut_args(|arg| arg.env(None::<&'static str>))
-}
+#[path = "common/pinned_cli.rs"]
+mod pinned_cli;
 
-/// Parse `argv` through the binary's own clap command with the environment
-/// pinned out, giving the clap defaults for everything argv omits.
-///
-/// Two drifts nothing below catches: a test here calling `Cli::parse_from`
-/// instead of this, and an env-backed arg on a future subcommand —
-/// `mut_args` and `get_arguments` both walk top-level args only.
-fn pinned_cli(argv: &[&str]) -> Cli {
-    let matches = pin_environment(bugwarden::config::command())
-        .try_get_matches_from(argv)
-        .expect("the harness command line must parse");
-    Cli::from_arg_matches(&matches).expect("the harness command line must build a Cli")
-}
+use pinned_cli::pinned;
 
 /// Build the http-transport `Cli` against `mock`, with `key_file` when the
 /// test runs in server-held mode.
 ///
-/// No clap `env` fallback reaches this `Cli`: [`pinned_cli`] drops them
+/// No clap `env` fallback reaches this `Cli`: [`pinned`] drops them
 /// wholesale rather than overwriting a list of named fields — the list is
 /// what drifted in #190, having lost `read_only` while still claiming to be
 /// complete. That is the whole of the claim; the rest of the environment
@@ -118,7 +97,7 @@ fn pinned_cli(argv: &[&str]) -> Cli {
 /// the only field a caller varies.
 fn http_cli(mock: &MockServer, key_file: Option<&std::path::Path>) -> Arc<Cli> {
     let uri = mock.uri();
-    let mut cli = pinned_cli(&[
+    let mut cli: Cli = pinned(&[
         "bugwarden",
         "--bugzilla-server",
         &uri,
@@ -131,72 +110,14 @@ fn http_cli(mock: &MockServer, key_file: Option<&std::path::Path>) -> Arc<Cli> {
     Arc::new(cli)
 }
 
-/// The pin is worth its doc comment only if the command really keeps no
-/// fallback, and only if the unpinned command still has some — a check over
-/// an already-empty set can never fail.
+/// The pin's own self-check, run in each binary that relies on it so a
+/// single-binary `cargo test --test ...` still proves what its harness
+/// claims. Both halves assert with their own message, so folding them into
+/// one test costs no diagnosis.
 #[test]
-fn the_pinned_command_keeps_no_environment_fallback() {
-    let mut pinned = pin_environment(bugwarden::config::command());
-    pinned.build();
-    let leaking: Vec<String> = pinned
-        .get_arguments()
-        .filter_map(clap::Arg::get_env)
-        .map(|env| env.to_string_lossy().into_owned())
-        .collect();
-    assert!(
-        leaking.is_empty(),
-        "these environment fallbacks still reach the harness: {leaking:?}"
-    );
-
-    let mut real = bugwarden::config::command();
-    real.build();
-    assert!(
-        real.get_arguments().any(|arg| arg.get_env().is_some()),
-        "the binary declares no environment fallback at all, so the assertion \
-         above proves nothing"
-    );
-}
-
-/// The drift the pin exists to survive: a `Cli` field added later, with an
-/// `env` fallback whose variable is set in the runner's environment. Any
-/// variable this process already carries reproduces it, so the test never
-/// mutates the environment libtest shares across threads.
-#[test]
-fn the_pin_neutralises_an_environment_backed_flag_added_later() {
-    let (name, value) = std::env::vars_os()
-        .next()
-        .expect("the test process must carry at least one environment variable");
-    // clap only accepts a `'static` variable name without its `string`
-    // feature; one leaked name per run is cheaper than turning that on.
-    let name: &'static std::ffi::OsStr = Box::leak(name.into_boxed_os_str());
-    let added_later = || {
-        clap::Arg::new("added-later")
-            .long("added-later")
-            .env(name)
-            .value_parser(clap::builder::OsStringValueParser::new())
-    };
-    let argv = ["bugwarden", "--bugzilla-server", "https://bugzilla.example"];
-
-    // Only the probe keeps a fallback: a malformed ambient value for a real
-    // one (`MCP_PORT=notanumber`) would otherwise fail this test.
-    let ambient = pin_environment(bugwarden::config::command())
-        .arg(added_later())
-        .try_get_matches_from(argv)
-        .expect("the probe command line must parse");
-    assert_eq!(
-        ambient.get_one::<std::ffi::OsString>("added-later"),
-        Some(&value),
-        "the probe variable must reach an unpinned field, or this proves nothing"
-    );
-
-    let pinned = pin_environment(bugwarden::config::command().arg(added_later()))
-        .try_get_matches_from(argv)
-        .expect("the probe command line must parse");
-    assert_eq!(
-        pinned.get_one::<std::ffi::OsString>("added-later"),
-        None,
-        "{name:?} reached a pinned field"
-    );
+fn the_environment_pin_holds() {
+    pinned_cli::assert_the_pin_drops_every_fallback::<Cli>();
+    pinned_cli::assert_the_pin_neutralises_a_flag_added_later::<Cli>();
 }
 
 /// Serve a [`BugWarden`] built from `cli` and `policy` over a real TCP
